Added Look Controller
The player can now look around the scene as they move. View movement is smoothed over time so that the view alignes with the axis input more like it would in real life. This makes it so that looking around is smooth and feels good on either a keyboard or a controller.